pyproject.toml
[project]
name = "llm-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-aiplatform>=1.111.0",
    "google-genai>=1.33.0",
    "google-generativeai>=0.8.5",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.6.6",
    "openai>=1.105.0",
]
